first commit - using ef_search=850, as the number of images of the biggest label.

+++ src/main/java/org/benaya/ai/parquet_to_postgres_triplets/model/LabelToEmbeddingId.java
@@ -12,6 +12,6 @@ import java.io.Serializable;
 @Setter
 @EqualsAndHashCode
 public class LabelToEmbeddingId implements Serializable {
-    private Integer label;
+    private Long label;
     private String path;
 }
